backend/requirements.txt
# ─────────────────────────────────────────────
# Cura3.ai Backend — Python Dependencies
# ─────────────────────────────────────────────

# ── Web Framework ────────────────────────────
fastapi>=0.115.0
uvicorn[standard]>=0.30.0

# ── MongoDB ──────────────────────────────────
motor>=3.6.0
pymongo>=4.9.0

# ── Auth & Security ──────────────────────────
python-jose[cryptography]==3.3.0
authlib==1.3.2
httpx>=0.28.1
itsdangerous==2.2.0

# ── AI / LLM ────────────────────────────────
langchain-core>=0.3.0
langchain-openai>=0.3.0
openai>=1.0.0

# ── Config ───────────────────────────────────
pydantic[email]>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0

# ── File Parsing ─────────────────────────────
PyPDF2>=3.0.0
python-docx>=1.1.0
python-multipart>=0.0.12

# ── PDF Generation ───────────────────────────
reportlab>=4.2.0

# ── Utilities ────────────────────────────────
# bson ObjectId is provided by pymongo (no separate install needed)

# ── Monitoring (Azure Application Insights) ──
opencensus-ext-azure>=1.1.0
